+++ crosshatch/Http402/server.ts
@@ -24,7 +24,7 @@ export const require = Effect.fnUntraced(function* ({ required }: { readonly req
   })
 })
 
-export const addSettlement =
+export const withSettlement =
   (settlement: typeof SettleResponse.Type) => (response: HttpServerResponse.HttpServerResponse) =>
     S.encodeEffect(SettleResponseFromBase64JsonString)(settlement).pipe(
       Effect.map((v) => HttpServerResponse.setHeader(PAYMENT_RESPONSE, v)(response)),
