+++ crosshatch/index.ts
@@ -19,7 +19,7 @@ export * as Facilitator from "./Facilitator/Facilitator.ts"
 export * as Extension from "./Extension.ts"
 export * as PaymentId from "./PaymentId.ts"
 export * as Mnemonic from "./Mnemonic.ts"
-export * as Http from "./Http/Http.ts"
+export * as Http402 from "./Http402/Http402.ts"
 export * as KnownAssets from "./KnownAssets/KnownAssets.ts"
 export * as Accept from "./Accept.ts"
 export * as Bridge from "./Bridge.ts"
