+++ facilitator/main.ts
@@ -1,5 +1,5 @@
 import { CredentialsFromEnv } from "@distilled.cloud/coinbase"
-import { Http, Facilitator } from "crosshatch"
+import { Facilitator } from "crosshatch"
 import { Layer, Effect } from "effect"
 import { Worker } from "effect-workerd"
 import { HttpRouter, HttpServerResponse } from "effect/unstable/http"
@@ -21,7 +21,6 @@ export default Worker.make({
             allowedHeaders: ["*"],
             allowedMethods: ["*"],
             allowedOrigins: ["*"],
-            exposedHeaders: Http.EXPOSED_HEADERS,
           }),
         ),
       ),
